requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django~=2.1.3
django-import-export==1.2.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.5.12
pytz==2018.9
PyYAML==3.13
requests==2.21.0
stripe==2.17.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
django-google-tag-manager